Refactor extension inheritance
